handling indicies with gauge masking

+++ torchdms/utils.py
@@ -241,4 +241,4 @@ def get_mutation_indicies(mutation_list, alphabet):
         site = int(mut[1:-1])
         indicies.append(((site - 1) * len(alphabet_dict)) + alphabet_dict[mut_aa])
 
-    return torch.Tensor(indicies)
+    return torch.Tensor(indicies).type(torch.long)
